Accept @font-face weight, width and oblique ranges and font-stretch percentages
A variable face covers its declared range in nearest-face matching, the weight, width and slant of the text set its wght, wdth and slnt axes inside the range, and nothing is faked that an axis supplies. A variable font with no descriptors covers the range of its own axes. The width is now a percentage end to end.

+++ src/PeachPDF.Tests/Html/Core/Utils/FontFamilyResolverPresentationTests.cs
@@ -38,10 +38,10 @@ namespace PeachPDF.Tests.Html.Core.Utils
 
             public List<EmojiPresentation> SystemFallbackRequests { get; } = [];
 
-            protected override RFont? CreateFontForCodepointInt(string family, double size, RFontStyle style, int weight, int stretch, double? obliqueSkewSinus, Rune codepoint, string? variations) =>
+            protected override RFont? CreateFontForCodepointInt(string family, double size, RFontStyle style, int weight, double stretch, double? obliqueSkewSinus, Rune codepoint, string? variations) =>
                 Families.GetValueOrDefault(family);
 
-            protected override RFont? CreateSystemFallbackFontForCodepointInt(double size, RFontStyle style, int weight, int stretch, double? obliqueSkewSinus, Rune codepoint, EmojiPresentation presentation, string? variations)
+            protected override RFont? CreateSystemFallbackFontForCodepointInt(double size, RFontStyle style, int weight, double stretch, double? obliqueSkewSinus, Rune codepoint, EmojiPresentation presentation, string? variations)
             {
                 SystemFallbackRequests.Add(presentation);
                 return SystemFallback.GetValueOrDefault(presentation);
